+++ activerecord/lib/active_record/connection_adapters/postgresql/schema_statements.rb
@@ -537,6 +537,13 @@ module ActiveRecord
               when 0..6; "timestamp(#{precision})"
               else raise(ActiveRecordError, "No timestamp type has precision of #{precision}. The allowed range of precision is from 0 to 6")
             end
+          when 'interval'
+            return super unless precision
+
+            case precision
+              when 0..6; "interval(#{precision})"
+              else raise(ActiveRecordError, "No interval type has precision of #{precision}. The allowed range of precision is from 0 to 6")
+            end
           else
             super
           end